+++ src/libs/auth/src/openid/types.rs
@@ -15,7 +15,7 @@ pub(crate) mod interface {
 }
 
 pub(crate) mod errors {
-    use crate::openid::jkwset::types::errors::GetOrRefreshJwksError;
+    use crate::openid::jwkset::types::errors::GetOrRefreshJwksError;
     use crate::openid::jwt::types::errors::{JwtFindProviderError, JwtVerifyError};
     use candid::{CandidType, Deserialize};
     use serde::Serialize;
