+++ src/observatory/src/lib.rs
@@ -16,7 +16,7 @@ use crate::types::interface::GetNotifications;
 use crate::types::interface::NotifyStatus;
 use crate::types::state::Env;
 use ic_cdk_macros::export_candid;
-use junobuild_auth::openid::jkwset::types::interface::GetOpenIdCertificateArgs;
+use junobuild_auth::openid::jwkset::types::interface::GetOpenIdCertificateArgs;
 use junobuild_auth::openid::types::provider::OpenIdCertificate;
 use junobuild_shared::types::interface::NotifyArgs;
 use junobuild_shared::types::interface::{DeleteControllersArgs, SetControllersArgs};
